Download
git-svn-id: https://compbio.charite.de/public-svn/NGSanalysis@1670 a7cbf6ab-7213-4657-a3ac-d80d95007299

+++ jannovar/src/main/java/jannovar/io/UCSCDownloader.java
@@ -19,7 +19,7 @@ public class UCSCDownloader {
     /** Path of direrctory to which the files will be downloaded. */
     private String directory_path;
     /** Base URI for UCSC hg19 build annotation files */
-    private String hg19base = "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/";
+    private String hg19base = "ftp://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/";
 
     /** Name of the knownGenes file. */
     private String knownGene = "knownGene.txt.gz";
@@ -49,6 +49,9 @@ public class UCSCDownloader {
      */
     public UCSCDownloader(String dirpath, String proxyHost, String port) {
 	this(dirpath);
+	if (proxyHost == null)
+	    return; /* Do not set proxy if proxyHost is null. */
+	System.setProperty("proxySet","true");
 	System.setProperty("http.proxyHost", proxyHost);
 	System.setProperty("http.proxyPort", port);
     }
@@ -61,23 +64,59 @@ public class UCSCDownloader {
      * warning message and skips it).
      */
     public void downloadUCSCfiles() {
+	makeDirectoryIfNotExist();
 	download_file(this.hg19base, this.knownGene);
+	download_file(this.hg19base, this.knownGeneMrna);
+	download_file(this.hg19base, this.kgXref);
+	download_file(this.hg19base, this.known2locus);
+	
 
     }
 
 
     /**
-     * This method downloads a file to the specified local file path
+     * This function creates a new directory to store the downloaded
+     * UCSC files. If the directory already exists, it just emits a
+     * warning and does nothing.
+     */
+    private void makeDirectoryIfNotExist() {
+	File directory = new File(this.directory_path);
+	if (directory.exists()) {
+	    System.err.println(String.format("Cowardly refusing to create "+
+					     "directory \"%s\" since it already exists",
+					     this.directory_path));
+	} else {
+	    directory.mkdir();
+	}
+    }
+
+
+    /**
+     * This method downloads a file to the specified local file path.
+     * If the file already exists, it emits a warning message and does nothing.
      */
     public boolean download_file(String baseURL, String fname ) {
 
 	String urlstring = baseURL + fname;
 	String local_file_path = this.directory_path + fname;
+	File f = new File(local_file_path);
+	if (f.exists()) {
+	    System.err.println(String.format("Timorously refusing to download "+
+					     "file \"%s\" since it already exists",
+					     local_file_path));
+	    return false;
 
+	}
+	System.err.println("Downloading: \"" + urlstring + "\"");
+	//System.out.println("File " + local_file_path);
+	//System.out.println("proxy: " +  System.getProperty("http.proxyHost"));
+	//System.out.println("port: " +  System.getProperty("http.proxyPort"));
+	int threshold = 0;
+	int block = 250000;
 	try{
 	    URL url = new URL(urlstring);
-	    url.openConnection();
-	    InputStream reader = url.openStream();
+	    
+	    InputStream reader = url.openConnection().getInputStream();
 	    FileOutputStream writer = new FileOutputStream(local_file_path);
 	    byte[] buffer = new byte[153600];
 	    int totalBytesRead = 0;
@@ -86,8 +125,13 @@ public class UCSCDownloader {
 		writer.write(buffer, 0, bytesRead);
 		buffer = new byte[153600];
 		totalBytesRead += bytesRead;
+		if (totalBytesRead > threshold) {
+		    System.err.print("=");
+		    threshold += block;
+		}
 	    }
-	    System.out.println("Done. " + (new Integer(totalBytesRead).toString()) + " bytes read.");
+	    System.err.println();
+	    System.err.println("Done. " + (new Integer(totalBytesRead).toString()) + " bytes read.");
 	    writer.close();
 	    reader.close();
 	} catch (MalformedURLException e){
